xsidebyside.com/requirements.txt
beautifulsoup4==4.12.2
email-validator==2.1.0.post1
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.25
trafilatura==1.6.1
Werkzeug==2.3.7